backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# GraphQL
strawberry-graphql==0.227.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0

# Authentication & Security
passlib==1.7.4
python-jose==3.5.0
PyJWT==2.10.1
cryptography==46.0.3
bcrypt==4.1.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Optional: Redis (for caching and task queue)
# redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.25.2

# Code Quality (Optional)
pylint==3.0.3
black==23.12.0
isort==5.13.2
